Adicionando Tratamento de Exceções

+++ src/main/java/com/newlevel/dscommerce/services/ProductService.java
@@ -2,14 +2,20 @@ package com.newlevel.dscommerce.services;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.newlevel.dscommerce.controllers.repositories.ProductRepository;
 import com.newlevel.dscommerce.dto.ProductDTO;
 import com.newlevel.dscommerce.entities.Product;
+import com.newlevel.dscommerce.services.exceptions.DataBaseException;
+import com.newlevel.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -20,7 +26,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product entity = productRepository.findById(id).get();
+		Product entity = productRepository.findById(id)
+				.orElseThrow(()-> new ResourceNotFoundException("Não foi possivel encontrar produto com id: "+ id));
 		return new ProductDTO(entity);
 	}
 	
@@ -43,16 +50,29 @@ public class ProductService {
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO productReceived) {
 		
-		Product product = productRepository.findById(id).get();
-		copyDtoToEntity(productReceived,product);
-		Product entitySaved =productRepository.save(product);
-		return new ProductDTO(entitySaved);
+		try {
+			Product product = productRepository.getReferenceById(id);
+			copyDtoToEntity(productReceived,product);
+			Product entitySaved =productRepository.save(product);
+			return new ProductDTO(entitySaved);
+			
+			
+		}catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Entidade não encontrada no banco de dados!");
+		}
 	
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);
+		if(!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Não foi possivel excluir, produto com id: "+ id + " não encontrado!");
+		}
+		try {
+			productRepository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("O sistema informa que a deleção não pode ser feita porque viola a integridade referencial dos dados");
+		}
 	}
